backend/pyproject.toml
[project]
name = "todolist"
version = "0.1.0"
description = ""
authors = [
    {name = "jerry.liu"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
